python/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wgrok"
version = "0.1.0"
description = "ngrok clone using Webex API as a message bus"
requires-python = ">=3.10"
dependencies = [
    "aiohttp>=3.9",
    "python-dotenv>=1.0",
    "webex-message-handler @ git+https://github.com/3rg0n/webex-message-handler#subdirectory=python",
]

[project.optional-dependencies]
dev = [
    "pytest>=7",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
    "ruff>=0.4",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests", "tests_e2e"]
